fix: Extract post helpers to their own modules to fix opengraph image generation.
